商户额度申请功能

+++ busi/src/main/java/com/welfare/service/AccountDepositApplyService.java
@@ -86,5 +86,5 @@ public interface AccountDepositApplyService {
    * @param id
    * @return
    */
-  AccountDepositApplyDetailInfo detail(Integer id);
+  AccountDepositApplyDetailInfo detail(Long id);
 }
